feat: test that askbirthday confirmation is valid
